macOS support, fixed WinCoords repr
