--- src/util.rs
@@ -1,16 +1,0 @@
-use windows::Win32::{Foundation::HINSTANCE, System::SystemServices::IMAGE_DOS_HEADER};
-
-pub fn get_instance_handle() -> HINSTANCE {
-    // Gets the instance handle by taking the address of the
-    // pseudo-variable created by the microsoft linker:
-    // https://devblogs.microsoft.com/oldnewthing/20041025-00/?p=37483
-
-    // This is preferred over GetModuleHandle(NULL) because it also works in DLLs:
-    // https://stackoverflow.com/questions/21718027/getmodulehandlenull-vs-hinstance
-
-    extern "C" {
-        static __ImageBase: IMAGE_DOS_HEADER;
-    }
-
-    unsafe { std::mem::transmute(&__ImageBase as *const _) }
-}
